+++ stable_cart/__init__.py
@@ -1,13 +1,12 @@
 """Public package exports for stable_cart."""
 
 from .evaluation import prediction_stability, accuracy
-from .less_greedy_tree import LessGreedyHybridRegressor, GreedyCARTExact
+from .less_greedy_tree import LessGreedyHybridRegressor
 
 __all__ = [
     "prediction_stability",
     "accuracy",
     "LessGreedyHybridRegressor",
-    "GreedyCARTExact",
 ]
 
 __version__ = "0.1.0"